app/requirements.txt
fastapi==0.101.1
pydantic==2.8.2 
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
asyncpg==0.30.0
vapi-python==0.1.0
vapi-server-sdk==1.7.2